requirements.txt
pymupdf>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0
openpyxl>=3.1.0
pyinstaller>=6.0.0
